services/reasoner/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "arc-sherlock"
version = "0.1.0"
description = "LangGraph reasoning engine for the A.R.C. platform"
requires-python = ">=3.13"
dependencies = [
    # Web framework
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "pydantic>=2.7",
    "pydantic-settings>=2.3",

    # LangGraph + LLM
    "langgraph>=1.0",
    "langchain-core>=0.3",
    "langchain-openai>=0.2",

    # Vector store + embeddings
    "sentence-transformers>=3.0",
    "pgvector>=0.3.0",

    # Relational storage
    "sqlalchemy>=2.0",
    "asyncpg>=0.30",
    "greenlet>=3.0",

    # Messaging
    "nats-py>=2.9",
    "pulsar-client>=3.4",

    # Observability
    "opentelemetry-api>=1.27",
    "opentelemetry-sdk>=1.27",
    "opentelemetry-instrumentation-fastapi>=0.48b0",
    "opentelemetry-instrumentation-sqlalchemy>=0.48b0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.27",
    "structlog>=24.4",

    # Resilience
    "tenacity>=9.0",

    # Chat API
    "langchain-anthropic>=0.3",
    "langchain-google-genai>=2.0",
    "sse-starlette>=2.1",
    "tiktoken>=0.7",

    # RAG
    "minio>=7.2",
    "python-multipart>=0.0.12",
    "pypdf>=4.0",
    "python-docx>=1.1",

    # Dev helpers (mounted only when SHERLOCK_DEV_MODE=true)
    "faker>=28.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.24",
    "httpx>=0.27",
    "pytest-cov>=5",
    "ruff>=0.6",
    "mypy>=1.11",
    "jsonschema>=4.23",
    "pyyaml>=6.0",
    # Type stubs
    "types-PyYAML",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.13"
strict = true
ignore_missing_imports = true
plugins = ["pydantic.mypy"]

# pulsar-client has no type stubs — union-attr errors on self._consumer / self._producer
# are false positives that can't be resolved without a typed wrapper around the C extension.
[[tool.mypy.overrides]]
module = ["sherlock.pulsar_handler"]
disable_error_code = ["union-attr"]

# minio SDK ships no type stubs — intentional suppression, same pattern as pulsar_handler above.
[[tool.mypy.overrides]]
module = ["sherlock.rag.adapters.minio"]
ignore_missing_imports = true

# FastAPI route decorators (router.post / router.get etc.) are untyped in the FastAPI stubs.
# Suppress untyped-decorator for all router modules — same false-positive as openai_router.
[[tool.mypy.overrides]]
module = [
    "sherlock.openai_router",
    "sherlock.models_router",
    "sherlock.files_router",
    "sherlock.vector_stores_router",
    "sherlock.embeddings_router",
]
disable_error_code = ["untyped-decorator"]
